[feature][vast][name] modify key name, modify err_t as error_t
[what] 修改key文件的命名，修改 err_t 为 error_t
[why]null
[how]null

+++ component/vast_store.c
@@ -529,7 +529,7 @@ static int16_t vast_store_writeToBackupBlock(Store_TypeTypeDef type, uint8_t *da
   * @param
   * @retval
   */
-int16_t vast_store_initialize(void)
+int vast_store_initialize(void)
 {
 	int16_t ret = VAST_OK;
 	uint32_t validAddr = 0xFFFFF, backupAddr = 0xFFFFF;
